refactor: :recycle: Refatorados os menus para padronização e organização
