+++ src/entities/menus/types.d.ts
@@ -1,0 +1,8 @@
+import { ReactNode } from 'react';
+
+interface Tab {
+    title: string;
+    icon: ReactNode;
+    type?: string;
+    path: string;
+}
